# Playground.Client/Generated/Models/Site.cs
// Code generated by Microsoft (R) AutoRest Code Generator 1.0.1.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace Playground.Client.Generated.Models
{
    using Newtonsoft.Json;
    using Playground.Client;
    using Playground.Client.Generated;
    using System.Linq;

    public partial class Site
    {
        /// <summary>
        /// Initializes a new instance of the Site class.
        /// </summary>
        public Site()
        {
          CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the Site class.
        /// </summary>
        public Site(int? id = default(int?), string name = default(string), int? companyId = default(int?), Company company = default(Company))
        {
            Id = id;
            Name = name;
            CompanyId = companyId;
            Company = company;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public int? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "name")]
        public string Name { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "companyId")]
        public int? CompanyId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "company")]
        public Company Company { get; set; }

    }
}