# Playground.Client/Generated/Models/SiteViewModel.cs
// Code generated by Microsoft (R) AutoRest Code Generator 1.0.1.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace Playground.Client.Generated.Models
{
    using Newtonsoft.Json;
    using Playground.Client;
    using Playground.Client.Generated;
    using System.Linq;

    public partial class SiteViewModel
    {
        /// <summary>
        /// Initializes a new instance of the SiteViewModel class.
        /// </summary>
        public SiteViewModel()
        {
          CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the SiteViewModel class.
        /// </summary>
        public SiteViewModel(string companyName = default(string), string siteName = default(string), string translation = default(string))
        {
            CompanyName = companyName;
            SiteName = siteName;
            Translation = translation;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "companyName")]
        public string CompanyName { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "siteName")]
        public string SiteName { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "translation")]
        public string Translation { get; set; }

    }
}