# Playground.Client/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "My API"
  },
  "basePath": "/",
  "paths": {
    "/Account/Login": {
      "post": {
        "tags": [ "Account" ],
        "operationId": "AccountLoginPost",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/Account/Register": {
      "post": {
        "tags": [ "Account" ],
        "operationId": "AccountRegisterPost",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/Account": {
      "post": {
        "tags": [ "Account" ],
        "operationId": "AccountPost",
        "consumes": [],
        "produces": [],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/connect/token": {
      "post": {
        "tags": [ "Account" ],
        "operationId": "ConnectTokenPost",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "AccessToken",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "AcrValues",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Assertion",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Claims",
            "in": "formData",
            "required": false,
            "type": "object"
          },
          {
            "name": "ClaimsLocales",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "ClientAssertion",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "ClientAssertionType",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "ClientId",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "ClientSecret",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Code",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "CodeChallenge",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "CodeChallengeMethod",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "CodeVerifier",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Display",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "GrantType",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "IdentityProvider",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "IdTokenHint",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "LoginHint",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "MaxAge",
            "in": "formData",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "Nonce",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Password",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "PostLogoutRedirectUri",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Prompt",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "RedirectUri",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "RefreshToken",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Request",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "RequestId",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "RequestUri",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Resource",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "ResponseMode",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "ResponseType",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Scope",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "State",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Token",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "TokenTypeHint",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Registration",
            "in": "formData",
            "required": false,
            "type": "object"
          },
          {
            "name": "UiLocales",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "Username",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/api/Companies": {
      "get": {
        "tags": [ "Companies" ],
        "operationId": "ApiCompaniesGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Company" }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": [ "Companies" ],
        "operationId": "ApiCompaniesPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Company" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Company" }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/Companies/GetViewModels": {
      "get": {
        "tags": [ "Companies" ],
        "operationId": "ApiCompaniesGetViewModelsGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/SiteViewModel" }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/Companies/{key}": {
      "get": {
        "tags": [ "Companies" ],
        "operationId": "ApiCompaniesByKeyGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Company" }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": [ "Companies" ],
        "operationId": "ApiCompaniesByKeyDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" }
        }
      },
      "patch": {
        "tags": [ "Companies" ],
        "operationId": "ApiCompaniesByKeyPatch",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "patch",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/JsonPatchDocument[Company]" }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Company" }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/Companies/{key}/Sites": {
      "get": {
        "tags": [ "Companies" ],
        "operationId": "ApiCompaniesByKeySitesGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Site" }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/Sites": {
      "get": {
        "tags": [ "Sites" ],
        "operationId": "ApiSitesGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Site" }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": [ "Sites" ],
        "operationId": "ApiSitesPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Site" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Site" }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/Sites/{key}": {
      "get": {
        "tags": [ "Sites" ],
        "operationId": "ApiSitesByKeyGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Site" }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": [ "Sites" ],
        "operationId": "ApiSitesByKeyDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" }
        }
      },
      "patch": {
        "tags": [ "Sites" ],
        "operationId": "ApiSitesByKeyPatch",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "patch",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/JsonPatchDocument[Site]" }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Site" }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/Values": {
      "get": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/Values/{id}": {
      "get": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "type": "string" }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "put": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "definitions": {
    "Company": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": { "type": "string" },
        "sites": {
          "type": "array",
          "items": { "$ref": "#/definitions/Site" }
        }
      }
    },
    "Site": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": { "type": "string" },
        "companyId": {
          "format": "int32",
          "type": "integer"
        },
        "company": { "$ref": "#/definitions/Company" }
      }
    },
    "SiteViewModel": {
      "type": "object",
      "properties": {
        "companyName": { "type": "string" },
        "siteName": { "type": "string" },
        "translation": { "type": "string" }
      }
    },
    "JsonPatchDocument[Company]": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": { "$ref": "#/definitions/Operation[Company]" },
          "readOnly": true
        }
      }
    },
    "Operation[Company]": {
      "type": "object",
      "properties": {
        "value": { "type": "object" },
        "path": { "type": "string" },
        "op": { "type": "string" },
        "from": { "type": "string" }
      }
    },
    "JsonPatchDocument[Site]": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": { "$ref": "#/definitions/Operation[Site]" },
          "readOnly": true
        }
      }
    },
    "Operation[Site]": {
      "type": "object",
      "properties": {
        "value": { "type": "object" },
        "path": { "type": "string" },
        "op": { "type": "string" },
        "from": { "type": "string" }
      }
    }
  },
  "securityDefinitions": {
    "default": {
      "flow": "password",
      "tokenUrl": "/connect/token",
      "type": "oauth2"
    }
  }
}